+++ hotel-service/src/main/java/com/hotel/dao/RoomTypeRepository.java
@@ -4,5 +4,5 @@ import com.hotel.entity.RoomType;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoomTypeRepository extends BaseRepository<RoomType, Long> {
+public interface RoomTypeRepository extends BaseRepository<RoomType, Integer> {
 }
