+++ hotel-service/src/main/java/com/hotel/service/RoomTypeService.java
@@ -5,5 +5,5 @@ import com.hotel.entity.RoomType;
 import org.springframework.stereotype.Service;
 
 @Service
-public class RoomTypeService extends BaseService<RoomType, Long, RoomTypeRepository> {
+public class RoomTypeService extends BaseService<RoomType, Integer, RoomTypeRepository> {
 }
